Make tests independent
+ Extract ServerSuite to prepare all stuff
